# app.js
require("dotenv").config();
const express = require("express");
const cors = require("cors");
const cookieParser = require("cookie-parser");
const bodyParser = require("body-parser");
const { OAuth2Client } = require("google-auth-library");

const app = express();
const port = process.env.PORT || 8080;

// Body-parser middleware
app.use(bodyParser.json());
app.use(bodyParser.urlencoded({ extended: true }));

// Middleware
app.use(cors());
app.use(express.json());
app.use(express.urlencoded({ extended: true }));
app.use(cookieParser());

// Google Authentication
const CLIENT_ID = process.env.GOOGLE_CLIENT_ID;
const CLIENT_SECRET = process.env.GOOGLE_CLIENT_SECRET;
const REDIRECT_URI = process.env.GOOGLE_REDIRECT_URI;
const oauth2Client = new OAuth2Client(CLIENT_ID, CLIENT_SECRET, REDIRECT_URI);

// Serve static files with the /public prefix
app.use("/public", express.static("public"));

// Routes
const authRoutes = require("./routes/authRoutes");
app.use("/auth", authRoutes);

// Route to start OAuth flow
app.get("/auth/google", (req, res) => {
  const authUrl = oauth2Client.generateAuthUrl({
    access_type: "offline",
    scope: ["profile", "email"],
  });
  res.redirect(authUrl);
});

// OAuth2 callback route
app.get("/oauth2callback", async (req, res) => {
  try {
    const { code } = req.query;
    const { tokens } = await oauth2Client.getToken(code);
    oauth2Client.setCredentials(tokens);

    // Log tokens to ensure they are being retrieved correctly
    console.log("Tokens received:", tokens);

    // Verify ID token
    const ticket = await oauth2Client.verifyIdToken({
      idToken: tokens.id_token,
      audience: CLIENT_ID,
    });
    const payload = ticket.getPayload();
    const email = payload.email;

    // Save tokens in a cookie
    res.cookie("authToken", tokens.id_token, { httpOnly: false });

    // Log the cookies to ensure they are being set correctly
    console.log("Cookies set:", req.cookies);

    // Redirect to googleLogin.html with email as URL parameter
    res.redirect(`/public/html/googleLogin.html?email=${email}`);
  } catch (error) {
    console.error("Error during OAuth2 callback", error);
    res.status(500).send("Authentication failed");
  }
});

// Route to set the authToken in local storage and redirect to index.html
app.get("/store-token", (req, res) => {
  res.send(`
    <script>
      localStorage.setItem('authToken', '${req.cookies.authToken}');
      window.location.href = '/public/html/index.html';
    </script>
  `);
});

// Serve the main index.html file
app.get("/", (req, res) => {
  res.sendFile("public/html/index.html", { root: "." });
});

// Google Bucket Credentials
const bucketName = process.env.BUCKET_NAME;
const keyFile = process.env.KEYFILENAME;

const { Storage } = require("@google-cloud/storage");
const storage = new Storage({ keyFilename: keyFile });
const googleBucketMiddleware = require("./middlewares/googleBucketMiddleware");

// Google Bucket
app.get("/image/:mediapath", async (req, res) => {
  try {
    const mediaPath = req.params.mediapath;
    const imageData = await googleBucketMiddleware.downloadIntoMemory(
      mediaPath
    );
    res.setHeader("Content-Type", "image/jpeg"); // Or the appropriate MIME type
    res.send(imageData[0]);
  } catch (err) {
    res.status(500).json({ error: "Failed to download image" });
  }
});

app.post("/image", googleBucketMiddleware.uploadFromMemory);
// const Busboy = require('busboy');

// app.post("/image", async (req, res) => {
//   try {
//     const busboy = Busboy({ headers: req.headers });

//     let fileBuffer;
//     let filename;
//     let mimetype;

//     busboy.on("file", (fieldname, file, filename, encoding, mimetype) => {
//       const chunks = [];
//       file.on("data", (chunk) => chunks.push(chunk));
//       file.on("end", () => {
//         fileBuffer = Buffer.concat(chunks);
//         console.log("File uploaded and buffered:", filename);
//       });
//     });

//     busboy.on("finish", async () => {
//       try {
//         const uploadedImage = await googleBucketMiddleware.uploadFromMemory(
//           fileBuffer
//         );
//         console.log("Image uploaded successfully:", uploadedImage);
//         res.status(200).json({ message: "Success" }); // Assuming you're using Express
//       } catch (err) {
//         console.error("Error uploading image to GCS:", err);
//         res.status(500).json({ error: err.message }); // Send a more specific error
//       }
//     });

//     // Pipe the request to Busboy
//     req.pipe(busboy);

//   } catch (err) {
//     console.error("Unexpected error during image upload:", err);
//     res.status(500).json({ error: 'An unexpected error occurred' });
//   }
// });

const dbConfig = require("./dbConfig");
const volunteerController = require("./controllers/volunteerController");
const organisationController = require("./controllers/organisationController");
const postController = require("./controllers/postController");
const listingController = require("./controllers/listingController");
const likesController = require("./controllers/likesController");
const searchPageController = require("./controllers/userSearchPageController");
const userFeedPageController = require("./controllers/userFeedPageController");
const userProfileController = require("./controllers/userProfileController");
const followController = require("./controllers/followController");
const signUpController = require("./controllers/signUpController");
const savedListingController = require("./controllers/savedListingController");
const commentController = require("./controllers/commentController");
const swaggerUi = require("swagger-ui-express");
const swaggerDocument = require("./swagger-output.json"); // Import generated spec

const sql = require("mssql");

app.use((req, res, next) => {
  res.setHeader("Access-Control-Allow-Origin", "null"); // Replace with your client origin
  res.setHeader("Access-Control-Allow-Methods", "GET"); // Add other methods if needed
  res.setHeader("Access-Control-Allow-Headers", "Content-Type"); // Add other allowed headers
  next();
});

app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));

app.get("/organisation/details/:id", organisationController.getOrgDetails);
app.get("/volunteers", volunteerController.getAllVolunteers); //get all user
app.get("/volunteers/:id", volunteerController.getVolunteerById); // Get user by ID
app.get("/volunteers/:username", volunteerController.getVolunteerByUsername);
app.put("/volunteers/:id", volunteerController.updateVolunteerProfile);
app.delete("/volunteers/:id", volunteerController.deleteVolunteer);

app.get("/organisations", organisationController.getAllOrganisations); //get all organisation
app.get("/organisations/:id", organisationController.getOrgById);
app.put("/organisations/:id", organisationController.updateOrgProfile);
app.delete("/organisations/:id", organisationController.deleteOrg);

app.get("/organisations/:OrgName", organisationController.getOrgByName);
//app.put("/organisations/:OrgName", organisationController.updateOrgProfile);

app.get("/listing", listingController.getAllListings);
app.get("/listing/byOrgId/:orgID", listingController.getListingsByOrgId);
app.get("/listing/byListingID/:id", listingController.getListingsByListingId);
app.get("/listing/:username", listingController.getListingByListingName);
app.post("/listing", listingController.postListing);

// Caden's Parts
app.get("/searchPage/allFollower/:id", searchPageController.getFollowersByID);
app.get(
  "/searchPage/allFollower",
  searchPageController.getAllFollowerRelations
);
app.post("/searchPage/postFollow", followController.postFollow);
app.delete("/searchPage/deleteFollow", followController.deleteFollow);
app.get("/searchPage", searchPageController.getAllAccounts);
app.get("/searchPage/:username", searchPageController.getAccountByUsername);

// Likes' end points
app.get("/likes/:id", likesController.getAllLikesById);
app.post("/likes", likesController.postLikeById);
app.delete("/likes", likesController.deleteLikesById);

app.get("/userFeedPage", userFeedPageController.getAllAccounts);
app.get("/comment/:id", commentController.getAllCommentsByPostID);
app.post("/userFeedPage", userFeedPageController.postComment);

// app.get("/userProfile/:id", postController.getAllPostsByAccID)
app.get("/userProfile/:id", volunteerController.getAllFollowersAndFollowing);
app.get("/volunteerProfile/:id", userProfileController.getAccountInfo);
app.get("/organisationProfile/:id", userProfileController.getOrganisationInfo);

app.get("/followedPost/:id", postController.getAllFollowedPosts);
app.post("/postCreation", postController.postPost);

app.get("/signUp/:AccID/:ListingID", signUpController.getAllSignUpByListingID);
app.post("/signUp", signUpController.postSignUp);
app.delete("/signUp", signUpController.deleteSignUp);

app.get(
  "/savedListing/:AccID/:ListingID",
  savedListingController.getAllSavedByListingID
);
app.post("/savedListing", savedListingController.postSaved);
app.delete("/savedListing", savedListingController.deleteSaved);

app.listen(port, async () => {
  try {
    // Connect to the database
    await sql.connect(dbConfig);
    console.log("Database connection established successfully");
  } catch (err) {
    console.error("Database connection error:", err);
    // Terminate the application with an error code (optional)
    process.exit(1); // Exit with code 1 indicating an error
  }

  console.log(`Server listening on port ${port}`);

  process.on("SIGINT", async () => {
    console.log("Server is gracefully shutting down");
    // Perform cleanup tasks (e.g., close database connections)
    await sql.close();
    console.log("Database connection closed");
    process.exit(0); // Exit with code 0 indicating successful shutdown
  });
});